memory reading to get sprint
reads peaks memory to get the current sprint value. by pressing the hotkey, G.
you start measure the value, it waits until G is pressed again.

then it displays the % used to make a climb.
this will be used for better communication between players so we can make it to the peak

todo:
add a GUI and custom hotkeys